chatbot-backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
openai>=1.52.0
python-dotenv>=1.0.0
pydantic>=2.7.4
mangum>=0.17.0
